Value::as_<number_type> works with smaller numbers
This means that as_u32 will return if underlying number was u32, u16, or
u8 as conversion to a bigger type is lossless. If you don't care about
the actual size and "just want a number", stick with the larger number
types like f64, u64 or i64.

Additionally: Stop generating Value type conversion functions and delete
that code generation script.
